requirements.txt
# Flask 相关
Flask>=2.3.2
Flask-CORS>=4.0.0
Werkzeug>=2.3.7
Jinja2>=3.1.2
itsdangerous>=2.1.2
click>=8.1.6

# 图像处理
Pillow>=9.0.0

# emoji 支持
emoji>=2.0.0

# 正则表达式
regex>=2023.10.3

# 日期时间处理
python-dateutil>=2.8.2

# 类型注解支持
typing-extensions>=4.8.0

# 其他常用库
requests>=2.25.0
python-dotenv>=1.0.0
gunicorn>=20.1.0

# 新增依赖
flask-limiter
apscheduler

# 新增依赖
markdown>=3.3.0
beautifulsoup4>=4.9.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5